+++ packages/drizzle/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @authhero/drizzle
 
+## 0.3.7
+
+### Minor Changes
+
+- Add resource servers, rules and permissions
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.79.0
+
 ## 0.3.6
 
 ### Patch Changes